+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Operation script `start_varlogsn.py` will not run the storage node if there are any missing data directories. (#VARLOG-765/#693)
 
 
 ## [0.1.14] - 2022-06-30